requirements.txt
streamlit
pillow
stripe
